fix(wec): refactor genre buttons
